backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.0.33
PyJWT==2.8.0
python-dotenv==1.0.0
